services/space-runtime/requirements.txt
# AWS Lambda runtime
mangum>=0.17.0

# Web framework (use older starlette to avoid pydantic v2)
fastapi==0.103.2
starlette==0.27.0

# HTTP client
httpx>=0.26.0
requests>=2.31.0
requests-toolbelt>=1.0.0

# Image processing
Pillow>=10.2.0

# Utilities
python-dotenv>=1.0.0
pydantic==1.10.13
